Makes consent completely optional and separates ConsentText component.
