+++ app/src/lib/components/shared/ConsentText/ConsentText.tsx
@@ -1,0 +1,30 @@
+import { Link } from "@mui/material";
+import React, { Fragment } from "react";
+
+export const CONSENT_ERROR_MESSAGE = "You must accept the terms and conditions of the sale.";
+
+export type ConsentType = "disclaimer" | "checkbox";
+
+export interface ConsentTextProps {
+  privacyHref?: string;
+  termsOfUseHref?: string;
+}
+
+export const ConsentText: React.FC<ConsentTextProps> = ({
+  privacyHref,
+  termsOfUseHref,
+}) => {
+  const linkElements = [
+    privacyHref ? <Link color="text.primary" href={ privacyHref } target="_blank">Privacy Notices</Link> : null,
+    termsOfUseHref ? <Link color="text.primary" href={ termsOfUseHref } target="_blank">Terms of Use</Link> : null,
+  ].filter(Boolean);
+
+  return (<>
+    have read, understood, and consent to the{" "}
+    { linkElements.map((linkElement, i) => {
+      return <Fragment key={ i }>{ i === linkElements.length - 1 ? "and " : "" }{ linkElement }{ " " }</Fragment>
+    }) }
+    of the sale.
+  </>);
+};
+
